use ptrdiff_t for pointer offsets

+++ magick/effect.c
@@ -1010,8 +1010,8 @@ static void Hull(const Image *image,const ssize_t x_offset,
 
   p=f+(columns+2);
   q=g+(columns+2);
-  r=q+(y_offset*((ssize_t) columns+2)+x_offset);
-  s=q-(y_offset*((ssize_t) columns+2)+x_offset);
+  r=q+(ptrdiff_t) (y_offset*((ssize_t) columns+2)+x_offset);
+  s=q-(ptrdiff_t) (y_offset*((ssize_t) columns+2)+x_offset);
 #if defined(MAGICKCORE_OPENMP_SUPPORT)
   #pragma omp parallel for schedule(static) \
     magick_number_threads(image,image,rows,2)
